+++ packages/machine-extractor/src/connect/modifySkyConfigSource.ts
