+++ packages/machine-extractor/src/connect/skyConfigNode.ts
