+++ packages/machine-extractor/src/index.ts
@@ -1,14 +1,13 @@
 export * from 'xstate/lib/types';
 export * from './MachineExtractResult';
+export * from './connect/extractSkyConfigFromFile';
+export * from './connect/modifySkyConfigSource';
+export * from './connect/skyConfigUtils';
 export * from './constants';
 export * from './extractMachinesFromFile';
 export * from './getMachineExtractResult';
 export * from './getMachineNodesFromFile';
 export * from './groupByUniqueName';
-export * from './skyConfig/createSkyConfigHelper';
-export * from './skyConfig/extractSkyConfigFromFile';
-export * from './skyConfig/modifySkyConfigSource';
-export * from './skyConfig/skyConfigUtils';
 export * from './stateNode';
 export * from './types';
 export * from './utils';
