+++ packages/shared/src/connect/fetchFromStately.ts
@@ -1,7 +1,7 @@
-import { SkyConfigFile } from '@xstate/tools-shared';
 import { AnyStateMachine } from 'xstate';
+import { SkyConfigFile } from './skyTypes';
 
-export const fetchFromStately = <T extends AnyStateMachine>(
+export function fetchFromStately<T extends AnyStateMachine>(
   {
     apiKey,
     url,
@@ -12,11 +12,11 @@ export const fetchFromStately = <T extends AnyStateMachine>(
     xstateVersion?: '4' | '5';
   },
   skyConfig?: SkyConfigFile<T>,
-) => {
+) {
   if (!skyConfig) {
     throw new Error(
       `You need to run xstate connect "src/**/*.ts?(x)" before you can use the Stately Sky actor with url ${url}`,
     );
   }
   return skyConfig.machine;
-};
+}
